# web/docs/swagger.yaml
definitions:
  api.addTagRequest:
    properties:
      result_id:
        type: integer
      tag_name:
        type: string
    type: object
...
  api.removeTagRequest:
    properties:
      result_id:
        type: integer
      tag_name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /results/tag/add:
    post:
      consumes:
      - application/json
      description: Associates a tag with a URL result. Creates the tag if it doesn't
        exist.
      parameters:
      - description: The result ID and tag name
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/api.addTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Add a tag to a result
      tags:
      - Tags
  /results/tag/remove:
    post:
      consumes:
      - application/json
      description: Removes the association between a tag and a URL result.
      parameters:
      - description: The result ID and tag name
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/api.removeTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Remove a tag from a result
      tags:
      - Tags
...
swagger: "2.0"